feat: add fiat estimation to split object
