+++ src/lib/objects/payggy/views/choose-amount.svelte
@@ -18,7 +18,7 @@
 	import Layout from '$lib/components/layout.svelte'
 	import { payggyDescriptor } from '..'
 	import type { ExchangeRateRecord } from '$lib/stores/exchangeRates'
-	import { getFiatAmountText } from '$lib/objects/utils'
+	import { getFiatAmountText } from '$lib/utils/fiat'
 
 	export let amount: string
 	export let token: TokenAmount
